Consolidate driver scripts
